refactor Board, Pieces, and Move for OOP
