+++ game/src/myCheckers/Board.java
@@ -4,14 +4,13 @@ import javafx.scene.Group;
 import javafx.scene.Parent;
 import javafx.scene.layout.Pane;
 
-import static myCheckers.TryMove.toBoard;
-import static myCheckers.TryMove.tryMove;
+import static myCheckers.Pieces.makePiece;
 
 public class Board
 {
     static final int TILE_SIZE = 100;
-    private static final int WIDTH = 8;
-    private static final int HEIGHT = 8;
+    static final int WIDTH = 8;
+    static final int HEIGHT = 8;
 
     public static Group tileGroup = new Group();
     public static Group pieceGroup = new Group();
@@ -52,58 +51,6 @@ public class Board
             }
         }
         return root;
-    }
-
-    public static Piece makePiece(PieceType type, int x, int y)
-    {
-        Piece piece = new Piece(type, x, y);
-
-        // Lambda method to get new layout and determine how to handle piece
-        piece.setOnMouseReleased(e -> {
-            int newX = toBoard(piece.getLayoutX());
-            int newY = toBoard(piece.getLayoutY());
-
-            MoveState result;
-
-            if (newX < 0 || newY < 0 || newX >= WIDTH || newY >= HEIGHT) {
-                // piece not moved so result is no move
-                result = new MoveState(MoveType.NONE);
-            } else {
-                // piece has attempted move, so gather data of attempted move
-                result = tryMove(piece, newX, newY);
-            }
-
-            // get the previous coordinates in order to clear the tile
-            int x0 = toBoard(piece.getOldX());
-            int y0 = toBoard(piece.getOldY());
-
-            // abort illegal move, or perform normal or capture (jump) move
-            switch (result.getType()) {
-                case NONE:
-                    // cancel illegal move
-                    piece.abortMove();
-                    break;
-                case NORMAL:
-                    // place piece in new position after move
-                    piece.move(newX, newY);
-                    board[x0][y0].setPiece(null);
-                    board[newX][newY].setPiece(piece);
-                    break;
-                case CAPTURE:
-                    // place piece in new position after capture
-                    piece.move(newX, newY);
-                    board[x0][y0].setPiece(null);
-                    board[newX][newY].setPiece(piece);
-
-                    // remove captured piece
-                    Piece otherPiece = result.getPiece(); // opponent's piece passed from TryMove
-                    board[toBoard(otherPiece.getOldX())][toBoard(otherPiece.getOldY())].setPiece(null);
-                    pieceGroup.getChildren().remove(otherPiece);
-                    break;
-            }
-        });
-
-        return piece;
     }
 
 }
